3 ghosts are done

+++ src/core/Core.cpp
@@ -65,7 +65,7 @@ void Core::mainLoop()
                     for (size_t x = 0; x < this->_menuMap[y].size(); x++)
                         this->_gfx->drawTile(this->_menuMap[y][x].tile, x, y);
 
-                int i = 0;
+                size_t i = 0;
                 rgb_t textColor = {123, 60, 0};
                 rgb_t selectedColor = {181, 49, 33};
 
@@ -74,7 +74,7 @@ void Core::mainLoop()
                 for (const auto &meta: this->_games) {
                     std::string name = meta.name;
                     name[0] = toupper(name[0]);
-                    
+
                     this->_gfx->drawTile(i == this->_selectedGame ? GAME_UNSELECTED : GAME_SELECTED, 7, 13 + i * 2);
                     this->_gfx->drawText(name, 9, 13 + i * 2, i == this->_selectedGame ? textColor : selectedColor);
                     i++;
